requirements.txt
fastapi
uvicorn
pillow
moviepy
requests
opencv-python
python-multipart
pydantic
